fix: 테스트 중복 코드 함수화
